+++ src/oolib/state/state_interface.nim
@@ -1,0 +1,8 @@
+import sugar
+import .. / classutil
+
+type
+  IState* = tuple
+    defConstructor: (
+      info: ClassInfo, partOfCtor: NimNode, argsList: seq[NimNode]
+    ) -> NimNode