+++ tests/isClass.nim
@@ -1,0 +1,19 @@
+discard """
+  action: "run"
+"""
+import ../src/oolib
+import unittest
+
+class A:
+  discard
+
+type B = ref object
+
+let
+  a {.used.} = new A
+  b {.used.} = new B
+
+check A.isClass()
+check a.isClass()
+check not B.isClass()
+check not b.isClass()